Initial Polish translation from Maciej Kobuszewski, properties files not yet updated to replace UTF-8 characters with \u escapes.
Potential fix to the problem of updateFolder sometimes failing. I'm
not certain, but I think that in at least some cases, even when it
throws an error the folder is still updated, so the error should just
be ignored. In any case, it's reasonable to ignore it since there's
nothing we can do about it.